Updated App.tsx
